Move msgbox_driver{,_ops} to a private header
This is a partial fix for issue: #171
"Remove static inline wrappers from driver headers"
