+++ drivers/msgbox/Makefile
@@ -3,4 +3,5 @@
 # SPDX-License-Identifier: BSD-3-Clause OR GPL-2.0-only
 #
 
+obj-y += msgbox.o
 obj-y += sunxi-msgbox.o
